Rename node.py to mix.py
