+++ txmix/__init__.py
@@ -6,7 +6,7 @@ for constructing mix networks with reduced code complexity
 from txmix.interfaces import IMixTransport, IRouteFactory
 from txmix.common import NodeDescriptor
 from txmix.client import ClientProtocol, MixClient, RandomRouteFactory
-from txmix.node import NodeProtocol, ThresholdMixNode
+from txmix.mix import NodeProtocol, ThresholdMixNode
 from txmix.udp_transport import UDPTransport
 from txmix.onion_transport import OnionTransport, OnionTransportFactory
 from txmix.utils import DummyPKI, RandReader, generate_node_keypair, generate_node_id, SphinxNodeKeyState
